fix(nh): actually set NH_FLAKE as well
I changed `up` to reference `NH_FLAKE` in
684923d2093571f8216fe35e6c0e948348e0e5c7, but was still only setting
`FLAKE`. nh was actually warning me about this, and I thought that the
fix would make that go away but it didn't. But now I got a hard error,
which finally helped me realize what my mistake was -- and this commit
fixes it